Continue move logic to check in store
