+++ src/stores/CheckInStore.ts
@@ -1,13 +1,16 @@
 import { makeAutoObservable, runInAction, toJS } from 'mobx';
 import rootStore from './RootStore';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+
+import { createCheckIn } from '@services/checkIn';
 import { Location, Event } from '@types/collections';
 
 class CheckInStore {
   rootStore: null | rootStore = null;
 
   pendingCheckIn: CheckIn | undefined = undefined;
-  activeCheckIn: CheckIn | undefined = undefined;
-  checkInPrivacy: PrivacyOptions | undefined = undefined;
+  lastCheckIn: CheckIn | undefined = undefined;
+  privacySetting: PrivacyOption | undefined = undefined;
   currentLocation: Location | undefined = undefined;
   currentEvent: Event | undefined = undefined;
 
@@ -31,9 +34,30 @@ class CheckInStore {
     this.activeCheckIn = checkInInfo;
   };
 
-  setCheckInPrivacy = (value: PrivacyOptions) => {
-    this.checkInPrivacy = value;
+  setPrivacySetting = (value: PrivacyOption) => {
+    this.privacySetting = value;
   };
+
+  async checkIn(checkInData: CheckIn) {
+    try {
+      const { checkIn } = await createCheckIn(checkInData);
+      this.lastCheckIn = checkIn;
+      await AsyncStorage.setItem('lastCheckIn', JSON.stringify(checkIn));
+      return checkIn;
+    } catch (err) {
+      console.error(err);
+      throw err;
+    }
+  }
+
+  async deleteLastCheckIn() {
+    try {
+      await AsyncStorage.removeItem('lastCheckIn');
+      this.lastCheckIn = undefined;
+    } catch (err) {
+      throw err;
+    }
+  }
 }
 
 export default CheckInStore;
